requirements.txt
torch
torchvision
Pillow
numpy
tqdm
matplotlib
open_clip_torch
git+https://github.com/openai/CLIP.git